Create first
this is a simple management tool for documents. i want to print combobox items with one click. this is an example. my code has over 100 documents. on some files error window is showing. whats wrong?  i think there is an easyer way to code that, but i have no clue how..